pyproject.toml
[project]
name = "najot-talim"
version = "0.1.0"
description = ""
authors = [
    {name = "Jace Trent",email = "trentjace58@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0,<7.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-stubs (>=3.16.6,<4.0.0)",
    "environs (>=14.5.0,<15.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode=false

